redo conditionals to be more readable
